Support set/get DRAM Max Throttle Enable commands
This commit supports Setting and Getting DRAM Max Throttle Enable via
ipmi.
Both Set and Get DRAM Max Throttle Enable commands are designed to be
flexible for each platform, users can configure the object paths of the
sensor in the "DRAM" property of "/user/ipmi-providers/power_limit.json"
file.
Each sensor has 2 properties "objectPath" and "requiredFlag". The
"objectPath" indicates the PLDM object path of this sensor. The
"requiredFlag" indicates whether this sensor is mandatory or not, if it
is true, BMC generates an error when BMC can not send the data to this
sensor, else BMC does not generate the error.
Example:
{
"DRAM":
{
  "pldm": [
  {
    "objectPath": "/xyz/openbmc_project/sensors/oem/S0_DRAM_MThrotEn",
    "requiredFlag": true
  },
  {
    "objectPath": "/xyz/openbmc_project/sensors/oem/S1_DRAM_MThrotEn",
    "requiredFlag": false
  }
  ]
}
}

Tested:
1. Set the DRAM Throttle Max Enable to true
   $ipmitool raw 0x3c 0x1e 0x01
2. Get the DRAM Throttle Max Enable status
   $ipmitool raw 0x3c 0x1f
3. Ipmitool return 0x01 (enable)
4. Set the DRAM Throttle Max Enable to false
   $ipmitool raw 0x3c 0x1e 0x00
5. Get the DRAM Throttle Max Enable status
   $ipmitool raw 0x3c 0x1f
6. Ipmitool return 0x00 (disable)

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>